mpm/requirements.txt
certifi==2020.11.8
chardet==3.0.4
idna==2.10
numpy==1.19.4
pandas==1.1.4
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
requests==2.25.0
six==1.15.0
urllib3==1.26.2
xlrd==1.2.0
